fix: update the existing ctx vars instead of creating a separate module
